fix ghl

+++ src/shared/usb/xsm3/excrypt_sha.c
@@ -9,10 +9,10 @@ void sha1_process_block(EXCRYPT_SHA_STATE* state)
 {
   uint32_t w[80];
   for (size_t i = 0; i < 16; i++) {
-    w[i] = (state->buffer[i * 4 + 0] << 24);
-    w[i] |= (state->buffer[i * 4 + 1] << 16);
-    w[i] |= (state->buffer[i * 4 + 2] << 8);
-    w[i] |= (state->buffer[i * 4 + 3]);
+    w[i] = ((uint32_t)state->buffer[i * 4 + 0] << 24);
+    w[i] |= ((uint32_t)state->buffer[i * 4 + 1] << 16);
+    w[i] |= ((uint32_t)state->buffer[i * 4 + 2] << 8);
+    w[i] |= ((uint32_t)state->buffer[i * 4 + 3]);
   }
   for (size_t i = 16; i < 80; i++) {
     w[i] = ROTL32((w[i - 3] ^ w[i - 8] ^ w[i - 14] ^ w[i - 16]), 1);
